requirements.txt
python-telegram-bot==22.7
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic-settings==2.11.0
structlog==24.1.0
alembic==1.13.1
pytest==8.0.0
httpx==0.27.2
SQLAlchemy==2.0.44
aiosqlite==0.20.0
psycopg[binary]==3.2.13
